pyproject.toml
[tool.poetry]
name = "drf-api"
version = "0.1.0"
description = ""
authors = ["OBADA ALHAWJREH <obada7jaber7@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
djangorestframework = "^3.12.4"
psycopg2-binary = "^2.9.1"
matplotlib = "^3.4.2"
Django = "^3.2.5"

[tool.poetry.dev-dependencies]
black = "^21.7b0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
